+++ app/src/main/java/at/bitfire/icsdroid/ui/EditCalendarActivity.kt
@@ -4,7 +4,6 @@
 
 package at.bitfire.icsdroid.ui
 
-import android.app.Application
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
@@ -18,23 +17,16 @@ import androidx.core.app.ShareCompat
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.FragmentTransaction
-import androidx.lifecycle.AndroidViewModel
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.viewModelScope
 import at.bitfire.icsdroid.HttpUtils
 import at.bitfire.icsdroid.R
-import at.bitfire.icsdroid.SyncWorker
 import at.bitfire.icsdroid.databinding.EditCalendarBinding
-import at.bitfire.icsdroid.db.AppDatabase
 import at.bitfire.icsdroid.db.dao.SubscriptionsDao
-import at.bitfire.icsdroid.db.entity.Credential
 import at.bitfire.icsdroid.model.CredentialsModel
+import at.bitfire.icsdroid.model.EditSubscriptionModel
 import at.bitfire.icsdroid.model.SubscriptionSettingsModel
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
 
 class EditCalendarActivity: AppCompatActivity() {
 
@@ -47,12 +39,12 @@ class EditCalendarActivity: AppCompatActivity() {
     private val subscriptionSettingsModel by viewModels<SubscriptionSettingsModel>()
     private val credentialsModel by viewModels<CredentialsModel>()
 
-    private val model by viewModels<SubscriptionModel> {
+    private val model by viewModels<EditSubscriptionModel> {
         object: ViewModelProvider.Factory {
             @Suppress("UNCHECKED_CAST")
             override fun <T : ViewModel> create(modelClass: Class<T>): T {
                 val subscriptionId = intent.getLongExtra(EXTRA_SUBSCRIPTION_ID, -1)
-                return SubscriptionModel(application, subscriptionId) as T
+                return EditSubscriptionModel(application, subscriptionId) as T
             }
         }
     }
@@ -221,79 +213,6 @@ class EditCalendarActivity: AppCompatActivity() {
 
     //TODO: Get rid of dirty model mechanism
     private fun dirty(): Boolean = true
-
-
-    /* view model and data source */
-
-    class SubscriptionModel(
-        application: Application,
-        private val subscriptionId: Long
-    ): AndroidViewModel(application) {
-
-        private val db = AppDatabase.getInstance(application)
-        private val credentialsDao = db.credentialsDao()
-        private val subscriptionsDao = db.subscriptionsDao()
-
-        val successMessage = MutableLiveData<String>()
-
-        val subscriptionWithCredential = db.subscriptionsDao().getWithCredentialsByIdLive(subscriptionId)
-
-        /**
-         * Updates the loaded subscription from the data provided by the view models.
-         */
-        fun updateSubscription(
-            subscriptionSettingsModel: SubscriptionSettingsModel,
-            credentialsModel: CredentialsModel
-        ) {
-            viewModelScope.launch(Dispatchers.IO) {
-                subscriptionWithCredential.value?.let { subscriptionWithCredentials ->
-                    val subscription = subscriptionWithCredentials.subscription
-
-                    val newSubscription = subscription.copy(
-                        displayName = subscriptionSettingsModel.title.value ?: subscription.displayName,
-                        color = subscriptionSettingsModel.color.value,
-                        defaultAlarmMinutes = subscriptionSettingsModel.defaultAlarmMinutes.value,
-                        defaultAllDayAlarmMinutes = subscriptionSettingsModel.defaultAllDayAlarmMinutes.value,
-                        ignoreEmbeddedAlerts = subscriptionSettingsModel.ignoreAlerts.value ?: false
-                    )
-                    subscriptionsDao.update(newSubscription)
-
-                    if (credentialsModel.requiresAuth.value == true) {
-                        val username = credentialsModel.username.value
-                        val password = credentialsModel.password.value
-                        if (username != null && password != null)
-                            credentialsDao.upsert(Credential(subscriptionId, username, password))
-                    } else
-                        credentialsDao.removeBySubscriptionId(subscriptionId)
-
-                    // notify UI about success
-                    successMessage.postValue(getApplication<Application>().getString(R.string.edit_calendar_saved))
-
-                    // sync the subscription to reflect the changes in the calendar provider
-                    SyncWorker.run(getApplication(), forceResync = true)
-                }
-            }
-        }
-
-        /**
-         * Removes the loaded subscription.
-         */
-        fun removeSubscription() {
-            viewModelScope.launch(Dispatchers.IO) {
-                subscriptionWithCredential.value?.let { subscriptionWithCredentials ->
-                    subscriptionsDao.delete(subscriptionWithCredentials.subscription)
-
-                    // sync the subscription to reflect the changes in the calendar provider
-                    SyncWorker.run(getApplication())
-
-                    // notify UI about success
-                    successMessage.postValue(getApplication<Application>().getString(R.string.edit_calendar_deleted))
-                }
-            }
-        }
-
-    }
-
 
     /** "Really delete?" dialog */
     class DeleteDialogFragment: DialogFragment() {
