+++ app/src/main/res/layout/edit_calendar.xml
@@ -1,6 +1,6 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android">
     <data>
-        <variable name="model" type="at.bitfire.icsdroid.ui.EditCalendarActivity.SubscriptionModel"/>
+        <variable name="model" type="at.bitfire.icsdroid.model.EditSubscriptionModel"/>
     </data>
 
     <ScrollView
